src/main/java/jdbcdao/utilities/ConnectionFactory.java
package jdbcdao.utilities;

import java.sql.Connection;
import java.sql.DriverManager;
import java.sql.SQLException;

public class ConnectionFactory {
    private String url;
    private String user;
    private String pass;

    public ConnectionFactory(String url, String user, String pass) {
        this.url = url;
        this.user = user;
        this.pass = pass;
    }

    public Connection getConnection()
    {
        Connection connection = null;
        try{
            connection = DriverManager.getConnection(url, user, pass);
            System.out.println("Connection Established to MYSQL Database");
        } catch (SQLException e) {
            System.err.println(e.getMessage());
            e.printStackTrace();
        }
        return connection;
    }
}